test(api): move POST /assets tests to test_assets.py to match assets.py
